数据导出修改

+++ database/migrations/2020_04_03_105332_create_members_table.php
@@ -32,7 +32,7 @@ class CreateMembersTable extends Migration
             $table->decimal('balance')->index()->default(0.00)->comment('余额');
             $table->tinyInteger('level')->index()->default(1)->comment('等级');
             $table->tinyInteger('status')->index()->default(1)->comment('状态');
-            $table->integer('referees')->index()->default(0)->comment('推荐人');
+            $table->unsignedBigInteger('referees')->index()->default(0)->comment('推荐人');
             $table->integer('active')->index()->default(30)->comment('活跃度');
             $table->double('longitude',10,3)->nullable()->comment('经度');
             $table->double('latitude',10,3)->nullable()->comment('纬度');
